Density analysis:
Estimate "curviness" of the toe.

Still need to come up with a better name then "strength" because when used for inversion, a lower "strength" provides a more pronounced visual effect

For stability's sake, first the base exponent is derived in a "straight" region of the density function, then afterwards the curve is fitted.

FIXME:  It really needs the ability for the user to specify the "straight" regions now.